Add files via upload
Assessing, Analyzing Wine data and Apply Machine Learning Algorithm that Classifies Quality level  
 Introduction: 

Wine is an alcoholic beverage made from fermented grapes. Yeast consumes the sugar in the grapes and converts it to ethanol, carbon dioxide, and heat.
It is a pleasant-tasting alcoholic beverage.
It will definitely be interesting to analyze the attributes of wine and understand their relationships and significance with wine quality label.